Update Total & qtt

+++ recap.php
@@ -42,9 +42,11 @@ if(!isset($_SESSION['products']) || empty($_SESSION['products'])){
             "</thead>",
             "<tbody>";
     $totalGeneral = 0;
-    $totalQtt= 0;
     foreach($_SESSION['products'] as $index=> $product){
-        echo "<tr>",
+
+        if($product['total']<$product['qtt']*$product['price']){
+            $product['total']=$product['qtt']*$product['price'];
+            echo "<tr>",
                 "<td class='p-3 mb-2 text-center'><strong>". $index ."</td>",
                 "<td class='p-3 mb-2 text-center '><strong><a class='text-decoration-none text-black' href='traitement.php?action=delete&id=".$index."'>". $product['name'] ."</a></td>",
                 "<td class='p-3 mb-2 text-center'><strong>".number_format($product['price'],2,",","&nbsp;")."&nbsp;€</td>",
@@ -55,7 +57,18 @@ if(!isset($_SESSION['products']) || empty($_SESSION['products'])){
                 "<td class='p-3 mb-2 text-center'><strong><a class='text-decoration-none' href='traitement.php?action=delete&id=".$index."'><button type='button' class='btn-close' aria-label='Close'></button></a></th>",
             "</tr>";
         $totalGeneral+= $product['total'];
-        $totalQtt += $product['qtt'];
+        }else{
+        echo "<tr>",
+                "<td class='p-3 mb-2 text-center'><strong>". $index ."</td>",
+                "<td class='p-3 mb-2 text-center '><strong><a class='text-decoration-none text-black' href='traitement.php?action=delete&id=".$index."'>". $product['name'] ."</a></td>",
+                "<td class='p-3 mb-2 text-center'><strong>".number_format($product['price'],2,",","&nbsp;")."&nbsp;€</td>",
+                "<td class='p-3 mb-2 text-center'><strong><i class='fa-regular fa-square-minus'></i></th>",
+                "<td class='p-3 mb-2 text-center'><strong>".$product['qtt'] ."</td>",
+                "<td class='p-3 mb-2 text-center'><strong><a class='text-decoration-none text-black' href='traitement.php?action=up-qtt&id=".$index."'><i class='fa-regular fa-square-plus'></i></th>",
+                "<td class='p-3 mb-2 text-center'><strong>".number_format($product['total'],2,",","&nbsp;")."&nbsp;€</td>",
+                "<td class='p-3 mb-2 text-center'><strong><a class='text-decoration-none' href='traitement.php?action=delete&id=".$index."'><button type='button' class='btn-close' aria-label='Close'></button></a></th>",
+            "</tr>";
+        $totalGeneral+= $product['total'];}
     }
     echo "<tr>",
             "<td colspan=6 class='text-center p-3 mb-2 bg-primary border border-white text-white'><strong>Total général : </td>",
